requirements.txt
bcrypt==3.2.2
cffi==1.15.0
click==7.1.2
mysql-connector-python==8.0.29
protobuf==4.21.1
pycparser==2.21
PyQt6==6.1.0
pyqt6-plugins==6.1.0.2.2
PyQt6-Qt6==6.3.0
PyQt6-sip==13.3.1
pyqt6-tools==6.1.0.3.2
python-dotenv==0.20.0
qt6-applications==6.1.0.2.2
qt6-tools==6.1.0.1.2
